validator: stop the self-configuring groups leaking their cluster
A few tests are about the cluster configuration itself, so they have
to build the node configs and start the cluster themselves. Give them
a fixture that starts nothing but stops whatever they started, once per
test rather than once per group, so a failure cannot leave a cluster
running into the next test.
